✨[Feat] 훅 분리
